v0.3.7-preview.6
This update merges in some big changes to the way the C# build scripts work! This allows for switching between the Oculus and Standard OpenXR Loader (for those _not_ targeting Quest!), as well as lays the groundwork for the upcoming Maui based multi-target template. Some projects _may_ experience issues upgrading due to this, especially if using the Universal template. If you see this in your project, I'd love to know, notes at the bottom about that.

Here's the changelog:
- Added auto-copy of Assets folder <SKCopyAssets>true/false, <SKAssetFolder>path, and <SKAssetDestination>path
- Added auto-compile for shaders in the Assets folder.
- Added ability to switch between the standard and Oculus version of the OpenXR Loader, <SKOpenXRLoader>.
- Added support for multimaterial skinned meshes, thanks @holloway! #495
- Added `Tex.GenParticle`.
- Added `Mesh.GenerateCircle`, and added `doubleSided` to `Mesh.GeneratePlane`, thanks @jackdaus! #503
- Added `Backend.OpenXR.OnPollEvent` to allow for more OpenXR extension implementations, thanks @jackdaus! #513
- Added `UI.Toggle` overload with sprite arguments. #514
- Added IntPtr overloads for `Sound.Read/WriteSamples`.
- Removed Linux dependencies on libjsoncpp and libglew. #507 and #485
- Text material now uses DepthTest.LessEq, should help reduce z-fighting artifacts on far text.
- Cmake now works with Android, thanks @Malek! #516
- Fixed an issue with a missing library on UWP ARM via cmake, #501
- Fixed an issue with parsing certain URIs in GLTF files.
- Fixed an issue with `Material.Copy` not copying material chaining.
- Fixed an issue with the simulated hand's wrist joint not matching real hand data.
- UI.Panel now uses the correct visual element. #484

Previously, asset copying was manually added to templates, this update adds some code that makes this no longer necessary. For existing projects, duplicate asset copying isn't a problem, but the Universal template did a bit extra in order to use a shared asset folder. Please see this branch of the Universal template for changes that you'll probably need!
https://github.com/StereoKit/SKTemplate-Universal/tree/feature/v0.3.7_Asset_Copy